# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      detail:
        description: 详细错误信息（可选）
        example: 详细错误信息
        type: string
      error:
        description: 错误信息
        example: 操作失败
        type: string
    type: object
...
  models.LoginRequest:
    properties:
      password:
        description: 密码
        example: admin
        type: string
      username:
        description: 用户名
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  models.LoginResponse:
    properties:
      access_token:
        description: Access Token
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      expires_in:
        description: Access Token 过期时间（秒）
        example: 604800
        type: integer
      refresh_token:
        description: Refresh Token
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token_type:
        description: Token 类型
        example: Bearer
        type: string
    type: object
...
  models.RefreshTokenRequest:
    properties:
      refresh_token:
        description: 刷新令牌
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  models.RevokeTokenRequest:
    properties:
      reason:
        description: 撤销原因
        example: 用户主动登出
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        description: 成功信息
        example: 操作成功
        type: string
    type: object
...
  services.RefreshResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
host: localhost:58091
info:
  contact:
    email: im.hanxi@gmail.com
    name: API Support
    url: https://github.com/songloft-org/songloft
  description: 轻量级音乐服务器 API 文档，支持本地音乐管理、网络歌曲、电台和歌单功能
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: Songloft API
  version: 2.12.1
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录获取访问令牌
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: 请求数据错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 用户登录
      tags:
      - 认证管理
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 用户登出，撤销当前访问令牌
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 用户登出
      tags:
      - 认证管理
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌
      parameters:
      - description: 刷新令牌请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功
          schema:
            $ref: '#/definitions/services.RefreshResponse'
        "400":
          description: 请求数据错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: 刷新令牌无效
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: 刷新令牌
      tags:
      - 认证管理
  /auth/tokens:
    get:
      consumes:
      - application/json
      description: 获取当前用户的所有活跃令牌列表
      parameters:
      - description: 令牌类型
        enum:
        - access
        - refresh
        in: query
        name: type
        type: string
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 令牌列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 列出活跃令牌
      tags:
      - 认证管理
  /auth/tokens/{token_id}:
    delete:
      consumes:
      - application/json
      description: 撤销指定的令牌
      parameters:
      - description: 令牌ID
        in: path
        name: token_id
        required: true
        type: string
      - description: 撤销令牌请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RevokeTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 撤销成功
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: 请求数据错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 撤销令牌
      tags:
      - 认证管理
...
securityDefinitions:
  BearerAuth:
    description: 输入 "Bearer {token}" 进行认证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"